cambios varios

+++ src/main/java/Config/Config.java
@@ -7,5 +7,5 @@ public class Config {
   public static final String URL_API = "https://apis.datos.gob.ar/georef/api/";
   public static final String RANKING_1 = "resources/ranking1.csv";
   public static final String RANKING_2 = "resources/ranking2.csv";
-  public static final int DISTANCIA_MINIMA = 100;
+  public static final double DISTANCIA_MINIMA = 100;
 }
